Refactoring Typo Business, upgrade to Wix install, update of VS solution file
