+++ WpfApplication1/App.cs
@@ -1,0 +1,53 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Globalization;
+using System.Threading;
+using WpfApplication1;
+
+public class App : Application
+{
+    readonly Window1 mainWindow;
+
+    // через систему внедрения зависимостей получаем объект главного окна
+    public App(Window1 mainWindow)
+    {
+        this.mainWindow = mainWindow;
+    }
+
+    protected override void OnStartup(StartupEventArgs e)
+    {
+        CultureInfo CI = (CultureInfo)Thread.CurrentThread.CurrentCulture.Clone();
+        CI.DateTimeFormat.ShortDatePattern = "dd.MM.yyyy";
+        Thread.CurrentThread.CurrentCulture = CI;
+        mainWindow.Show();
+        base.OnStartup(e);
+     }
+    public static void ChangeCulture(CultureInfo culture)
+    {
+        CultureInfo CI = (CultureInfo)culture.Clone();
+        CI.DateTimeFormat.ShortDatePattern = "dd.MM.yyyy";
+        Thread.CurrentThread.CurrentCulture = CI;
+        Thread.CurrentThread.CurrentUICulture = CI;
+        var oldWindow = Application.Current.MainWindow;
+
+// Calculation of Display Resolution
+//PresentationSource MainWindowPresentationSource = PresentationSource.FromVisual(oldWindow);
+//Matrix m = MainWindowPresentationSource.CompositionTarget.TransformToDevice;
+//// DpiWidthFactor = m.M11;
+//// DpiHeightFactor = m.M22;
+//double ScreenHeight = SystemParameters.PrimaryScreenHeight * m.M22;
+//double ScreenWidth = SystemParameters.PrimaryScreenWidth * m.M11;
+
+        Application.Current.MainWindow =  new Window1();
+        Application.Current.MainWindow.Show();
+        if (!Application.Current.MainWindow.ShowActivated)
+            Application.Current.MainWindow.Activate();
+        Application.Current.MainWindow.HorizontalAlignment = HorizontalAlignment.Stretch;
+        oldWindow.Close();
+
+    }
+}
