--- WpfApplication1/BuisenessLogic/IBuisenessLogic.cs
@@ -1,16 +1,0 @@
-﻿using WpfApplication1.DTO;
-
-namespace WpfApplication1
-{
-    interface IBuisenessLogic
-    {
-        void UpdateData();
-        void UpdateDataModel();
-        void FilterData();
-        void UpdateViewData();
-        void FinalizeBL();
-       // FilterViewModel Filter { get; set; }
-        DataRequest Request { get; }
-        ResponseModel ResponseModel { get; }
-    }
-}
