--- WpfApplication1/BuisenessLogic/SettingsLogic.cs
@@ -1,8 +1,0 @@
-﻿namespace WpfApplication1.BuisenessLogic
-{
-    class SettingsLogic
-    {
-        //Wrapper Methods for SettingsPresenter to access and modify Settings.Default  ...
-        // TO DO if necessary
-    }
-}
