+++ WpfApplication1/BusinessLogic/SettingsLogic.cs
@@ -1,0 +1,8 @@
+namespace WpfApplication1.BusinessLogic
+{
+    class SettingsLogic
+    {
+        //Wrapper Methods for SettingsPresenter to access and modify Settings.Default  ...
+        // TO DO if necessary
+    }
+}
